refactor(core): lift narrow_multi_segment to Provider trait default
The macOS (ax.rs) and Windows (uia.rs) providers each carried a
byte-identical copy of narrow_multi_segment (modulo three import-path
differences). The function only uses Provider trait methods
(get_children, find_elements) and core selector types, so it belongs
on the trait as a default method.

Net −48 lines, no behavior change. Linux already uses the default
find_elements path and never had a local narrow_multi_segment.
