+++ xa11y-macos/src/ax.rs
@@ -1037,7 +1037,7 @@ fn snake_to_ax_pascal(snake: &str) -> String {
 
 // ── Lightweight selector matching (no ElementData) ───────────────────────────
 
-use xa11y_core::selector::{match_op, Combinator, SimpleSelector};
+use xa11y_core::selector::{match_op, SimpleSelector};
 
 /// Test whether a raw AXElement matches a SimpleSelector, fetching only the
 /// attributes the selector actually inspects. This avoids building a full
@@ -1722,65 +1722,6 @@ impl MacOSProvider {
         results
     }
 
-    /// Narrow candidates through remaining selector segments (Child/Descendant
-    /// combinators), deduplicate, apply final :nth and limit.
-    fn narrow_multi_segment(
-        &self,
-        mut candidates: Vec<ElementData>,
-        segments: &[xa11y_core::selector::SelectorSegment],
-        max_depth: u32,
-        limit: Option<usize>,
-    ) -> Result<Vec<ElementData>> {
-        for segment in segments {
-            let mut next_candidates = Vec::new();
-            for candidate in &candidates {
-                match segment.combinator {
-                    Combinator::Child => {
-                        let children = self.get_children(Some(candidate))?;
-                        for child in children {
-                            if xa11y_core::selector::matches_simple(&child, &segment.simple) {
-                                next_candidates.push(child);
-                            }
-                        }
-                    }
-                    Combinator::Descendant => {
-                        let sub_selector = Selector {
-                            segments: vec![xa11y_core::selector::SelectorSegment {
-                                combinator: Combinator::Root,
-                                simple: segment.simple.clone(),
-                            }],
-                        };
-                        let mut sub_results = self.find_elements(
-                            Some(candidate),
-                            &sub_selector,
-                            None,
-                            Some(max_depth),
-                        )?;
-                        next_candidates.append(&mut sub_results);
-                    }
-                    Combinator::Root => unreachable!(),
-                }
-            }
-            let mut seen = HashSet::new();
-            next_candidates.retain(|e| seen.insert(e.handle));
-            candidates = next_candidates;
-        }
-
-        // Apply :nth on last segment
-        if let Some(nth) = segments.last().and_then(|s| s.simple.nth) {
-            if nth <= candidates.len() {
-                candidates = vec![candidates.remove(nth - 1)];
-            } else {
-                candidates.clear();
-            }
-        }
-
-        if let Some(limit) = limit {
-            candidates.truncate(limit);
-        }
-
-        Ok(candidates)
-    }
 }
 
 impl Provider for MacOSProvider {
